INF-1161: rewrite SDK paths to our new CDN
The SDK release artefacts are now published to a CDN ensuring fast
downloads for our users. See: https://github.com/dfinity-lab/sdk/pull/492

To make use of this CDN the sdk.dfinity.org site needs to
redirect (301) to the CDN at https://download.dfinity.systems.
This commit configures the following redirects:

* https://sdk.dfinity.org/install.sh → https://download.dfinity.systems/sdk/install.sh

* https://sdk.dfinity.org/manifest.json → https://download.dfinity.systems/sdk/manifest.json

* https://sdk.dfinity.org/downloads/dfx/* → https://download.dfinity.systems/sdk/dfx/*

Note that it's important not to use the 200 code since that will
instruct Netlify to proxy instead of redirect which will defeat the
purpose of the CDN.